requirements.txt
pixivpy
Pybooru
python-xmp-toolkit
pyyaml
yaql